backend/requirements.txt
# Django Core
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-extensions==3.2.3

# Database
# psycopg2-binary==2.9.9  # Commented out - using SQLite for development
dj-database-url==2.1.0

# Authentication
django-allauth==0.57.0
djangorestframework-simplejwt==5.3.0
python-jose[cryptography]==3.3.0

# Environment & Configuration
python-decouple==3.8
python-dotenv==1.0.1

# File Storage
# django-storages==1.14.2  # Commented out for development
# boto3==1.34.34  # Commented out for development
Pillow==10.2.0
# cloudinary==1.38.0  # Commented out for development
# django-cloudinary-storage==0.3.0  # Commented out for development

# Phone Numbers
django-phonenumber-field==7.3.0
phonenumbers==8.13.27

# API Documentation
drf-spectacular==0.27.1

# Async & Caching
# celery==5.3.4  # Commented out for development
# redis==5.0.1  # Commented out for development
# django-redis==5.4.0  # Commented out for development

# Utilities
python-slugify==8.0.2
django-import-export==3.3.5
openpyxl==3.1.2

# Development
django-debug-toolbar==4.2.0
factory-boy==3.3.0
Faker==22.2.0

# Testing
pytest==7.4.4
pytest-django==4.8.0
pytest-cov==4.1.0
pytest-xdist==3.5.0

# Code Quality
flake8==7.0.0
black==24.1.1
isort==5.13.2

# Monitoring
sentry-sdk[django]==1.40.0

# HTTP Client
requests==2.31.0
httpx==0.26.0

# Payment Gateways
razorpay==1.4.1

# Google Places API
googlemaps==4.10.0

# Date/Time
python-dateutil==2.8.2
pytz==2023.4

# Text Processing
markdown==3.5.2
bleach==6.1.0

# Deployment
gunicorn==21.2.0
whitenoise==6.6.0

# Development Server
django-livereload-server==0.4